+++ lib/central_config.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'active_support/core_ext/module/delegation.rb'
+require 'active_support/core_ext/module/delegation'
 
 require 'central_config/version'
 require 'central_config/engine' if defined?(Rails)
@@ -15,16 +15,22 @@ module CentralConfig
   module Base
     delegate :flag?, :load, :setting, to: :config
 
+    @@configure_block = proc {}
+
     def config
-      Thread.current[:central_config] ||= ::CentralConfig::Config.new
+      Thread.current[:central_config] ||= begin
+        conf = CentralConfig::Config.new
+        @@configure_block.call(conf)
+        conf
+      end
     end
 
     def config=(value)
       Thread.current[:central_config] = value
     end
 
-    def configure
-      yield config
+    def configure(&block)
+      @@configure_block = block
     end
   end